fix: use float32 to avoid dtype mismatch on T4 GPU
DeepSeek-OCR-2's infer() hardcodes autocast to bfloat16, which T4
doesn't support. With float16 weights the vision encoder outputs
float32 under autocast, causing masked_scatter_ dtype mismatch.
Loading as float32 keeps all tensors consistent.

+++ src/config.py
@@ -11,7 +11,7 @@ class ModelConfig:
     """Configuration for the DeepSeek OCR model."""
 
     model_id: str = "deepseek-ai/DeepSeek-OCR-2"
-    torch_dtype: str = "float16"
+    torch_dtype: str = "float32"
     base_size: int = 1024
     image_size: int = 768
     crop_mode: bool = True
